Update patch set 2
Patch Set 2:

Did you intentionally suggest to ship it? (Because you change to final release version) 

But, actually, I think it is okay to ship, it was out there for a while in "incorrect" alpha state and I didn't here about any issues related to it.

Let's see what Yigit thinks as well

Patch-set: 2