Update patch set 4
Patch Set 4:

Please resolved the build error:

/usr/local/google/buildbot/src/android/androidx-master-dev/frameworks/support/media2/src/main/java/androidx/media2/MediaSessionImplBase.java:184: Error: Access to private member of class MediaButtonReceiver requires synthetic accessor [SyntheticAccessor]
mBroadcastReceiver = new MediaButtonReceiver();
~~~~~~~~~~~~~~~~~~~~~~~~~

Explanation for issues of type "SyntheticAccessor":
A private inner class which is accessed from the outer class will force the
compiler to insert a synthetic accessor; this means that you are causing
extra overhead. This is not important in small projects, but is important
for large apps running up against the 64K method handle limit, and
especially for libraries where you want to make sure your library is as
small as possible for the cases where your library is used in an app
running up against the 64K limit.

Patch-set: 4
Reviewer: Gerrit User 1112679 <1112679@85c56323-6fa9-3386-8a01-6480fb634889>