Update patch set 2
Patch Set 2:

(2 comments)

General question: what promoted the decision to tie assets and resources together? It seems reasonable that developers might want to choose specifically one or the other. And this opens up the possibility that a developer sets one but not the other, and unexpectedly exposes the latter via the default domain and path.

Patch-set: 2
Reviewer: Gerrit User 1101490 <1101490@85c56323-6fa9-3386-8a01-6480fb634889>