Update patch set 4
Patch Set 4:

(7 comments)

> Patch Set 3:
> 
> (8 comments)
> 
> > Patch Set 2:
> > 
> > (2 comments)
> > 
> > General question: what promoted the decision to tie assets and resources together? It seems reasonable that developers might want to choose specifically one or the other. And this opens up the possibility that a developer sets one but not the other, and unexpectedly exposes the latter via the default domain and path.
> 
> Well, it was just because we couldn't see why the developer would want to explicitly disable one over the another. I mean the developer should be the only one who can load this local assets/resources so I don't see what the problem of always exposing them here.
> 
> However, If we want to enforce this, we can add two other methods to the builder class:
> allowAssetsHosting(boolean), allowResourcesHosting(boolean). But I'm not sure what the default values should be. (If we want to prevent unexpected exposure, then they should be disabled by default .. if we want to just add the ability of not exposing one of them, then they should be allowed.)

I don't think it's particularly interesting to allow one of them to not be exposed; they are both exposed in the current file:// implementation, and it doesn't seem likely that there is a meaningful distinction regarding security/etc between the assets and resources in a single APK.

I don't think exposing them on different domains is particularly useful, though, and that any real usage is likely to set them both to the same thing. Maybe we should just have a single method for that, i.e. just one choice of domain but still allow separate specified paths.

In general I think this API is good, but I have a bunch of style comments about the code below:

Patch-set: 4
Reviewer: Gerrit User 1021004 <1021004@85c56323-6fa9-3386-8a01-6480fb634889>