Fix spinner widget scroll
The bug caused a weird scrolling problem in the spinner widget when
used in dropdown mode. The problem is older, possibly introduced in
ag/680812, but as we still used the platform spinner for new API
versions, this was hidden. It came to light after aosp/789994, this
backported the dialog mode and uses the androidx version always.

The steps :
- try to scroll
- OnTouch from ListPopupWindow is called
- onTouch uses a ResizePopupRunnable which calls show() from
DropdownPopup (because it's overridden)
- show(int, int) from DropdownPopup is called which recreates
the popup

This works correctly for the spinner in dialog mode because the
show() method from ListPopupWindow is not overriden by the spinner

The fix is to delete the show() override in DropdownPopup which
makes the appcompat implementation of spinner work like in
platform.

Bug: 124274573
Test: ./gradlew appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.AppCompatSpinnerTest
Change-Id: I42aff5a5c8cb1417376dada2e3555d50d5e87164