Update patch set 2
Patch Set 2:

> ^ In that case, maybe LIBRARY_GROUP is the right thing for us? Or could you recommend another @RestrictTo value?

*meant "LIBRARY_GROUP_PREFIX," although maybe LIBRARY_GROUP could also be appropriate?

Also, there are should-always-be-hidden methods as well as soon-to-be-unhidden APIs. We could (and perhaps should) use different visibilities for these (this would hold us accountable to only call soon-to-be-unhidden APIs in our test app, not to cheat by using should-always-be-hidden methods).

Patch-set: 2
Reviewer: Gerrit User 1141809 <1141809@85c56323-6fa9-3386-8a01-6480fb634889>