This is another fix for the accessibility delegate probelem found in aosp/895295, which was reverted in aosp/913093.
Now the item delegate keeps a references to the app delegate. We add
either the item delegate or app delegate on bind and keep track of the
app delegate on recyclying. The delegate is never set to null whcih
prevents ViewCompat from adding its default implementation. If a
developer sets it to null the item delegate is added. Two RVs can share
views and maintain their own delegates with this tracking scheme.

Test: Tested with RecyclerViewAccessibilityTest and
RecyclerViewAccessibilityLifecycleTest. Tested with ViewPager-2 demos to
make sure the CollectionInfo was being updated correctly.

Change-Id: I9b26013515208e997f6a01eb26d282c3759da637