Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > API Council will have to review the transition library before this goes to beta.
> > 
> > There were two API changes both already reviewed and the feedback is addressed: b/120504681 and b/119620689.
> > Is it enough or something else should be done?
> 
> Nope.  Just a double check that we've abided by API council.  I'll +1 API review once presubmit succeeds.

Looks like it is not that easy to switch to use the stable versions of dependencies. Presubmit is failing because of the lint. Here we use some methods from core and fragment libs marked with LIBRARY_GROUP_PREFIX. But if we would use a stable version of dependency this methods would be still marked with just LIBRARY_GROUP there. So the current version of lint doesn't allow such usages.
Or we need to release everything together(which means no beta for transitions, right?), or release a special stable version for this dependencies with just an annotation updated or suppress this lint errors?
Nick, what do you think?

Patch-set: 1
Reviewer: Gerrit User 1361834 <1361834@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0