This is another fix for the accessibility delegate probelem found in aosp/895295, which was reverted in aosp/913093.
This adds a hidden function in AccessibilityDelegateCompat that checks if
the instance is a default delegateCompat by matching classname
or contains a null delegate. RV replaces the default delegate with
the item delegate with this check.

Test: Tested with RecyclerViewAccessibilityTest and
RecyclerViewAccessibilityLifecycleTest. Tested with ViewPager-2 demos to
make sure the CollectionInfo was being updated correctly.

Change-Id: I4549a577cfc3c3aa45c569b0673833c8eabb505c