Update patch set 2
Patch Set 2:

Thanks for your contribution.

As per the README.md (https://android.googlesource.com/platform/frameworks/support/+/androidx-master-dev#accepted-types-of-contributions), each contribution needs to be tied to an issue on https://issuetracker.google.com and feature work, such as this, must be approved by an AndroidX team member. Please file a bug with your use case before continuing with this CL.

Also note that every CL is expected to come with tests for the functionality they change.

Patch-set: 2
CC: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>