Update patch set 11
Patch Set 11:

I'm a bit confused by this request. The values of the constants are not meant to be used by developers. Their values come from the boundary interface definition, and they ensure that webview and androidx agree on the definition of the boundary interface they refer to. In this case, it is guaranteeing that the androidx API doesn't treat earlier released versions of webview as supporting an API that we had to change the signature of.

Patch-set: 11
Tag: mailMessageId=<CADFGwtg476F=7Uj7RKTUAMjomuW3gu1G94BDGYUx6P+PffH-ZQ@mail.gmail.com>