Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> I'm a bit confused by this request. The values of the constants are not meant to be used by developers. Their values come from the boundary interface definition, and they ensure that webview and androidx agree on the definition of the boundary interface they refer to. In this case, it is guaranteeing that the androidx API doesn't treat earlier released versions of webview as supporting an API that we had to change the signature of.
Value is something that is seen in javadocs and something that cannot change. It seems a bit odd to start shipping a brand new API (as far as public api goes) with 3 in it.

I assume you can use a constant without a 3 in public api, but then swap to a different constant (with a 3) when you get a call that then you can pass to webview/chromium.

Is that not feasible?

Patch-set: 11
Reviewer: Gerrit User 1086749 <1086749@85c56323-6fa9-3386-8a01-6480fb634889>