Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > > Patch Set 11:
> > > 
> > > I'm a bit confused by this request. The values of the constants are not meant to be used by developers. Their values come from the boundary interface definition, and they ensure that webview and androidx agree on the definition of the boundary interface they refer to. In this case, it is guaranteeing that the androidx API doesn't treat earlier released versions of webview as supporting an API that we had to change the signature of.
> > Value is something that is seen in javadocs and something that cannot change. It seems a bit odd to start shipping a brand new API (as far as public api goes) with 3 in it.
> > 
> > I assume you can use a constant without a 3 in public api, but then swap to a different constant (with a 3) when you get a call that then you can pass to webview/chromium.
> > 
> > Is that not feasible?
> 
> This is feasible, but requires some work. Right now, WebViewFeature.isFeatureSupported assumes WebViewFeature.PROXY_OVERRIDE (public API) has the same value as Features.PROXY_OVERRIDE (internal API).
> 
> Right now, WebViewFeature, WebViewFeatureInternal, and Features are all tightly coupled. We could decouple this by adding a new constructor for WebViewFeatureInternal, which accepts 2 strings: the internal value and the external value.
> 
> I agree with Aurimas: the ":3" is an implementation detail, and it's not appropriate to expose this, even if we expect apps to use named constants.

Sure, but these feature strings' values could just as easily be random, or they could be integer constants. The whole thing is an implementation detail underlying the constant names.

Placing it in the boundary interface means that the feature support is tightly coupled with the definition of the API. Keeping that constant while changing the interface, and updating a sidecar version code in two places in two different repositories seems more error prone.

I also don't believe that exposing the version to the developer is the wrong thing. For instance it's not clear in the case of an update to an API that we could express that WebView supported FEATURE:n and FEATURE:n+1 simultaneously, if we're mapping from external FEATURE, without implying that supporting version n+1 implies supporting version n (which isn't true, as evidenced in this case).

Does @WebViewSupportFeature support passing strings with the same value as those in the @StringDef, or do they have to have object identity?

Patch-set: 11