Add offscreenPageLimit to ViewPager2
Uses calculateExtraLayoutSpace to add pages that are off the screen to
the view hierarchy. Caches that keep items bound to views are disabled
to guarantee that pages are unbound once they are further then the
offscreen page limit away from the viewport.

Added OffscreenPageLimitTest to verify that the number of pages on
either side of the current viewport corresponds to the
offscreenPageLimit at all times.

Also added a new method requestTransform(), similar to requestLayout(),
that triggers the registered PageTransformer on demand.

Test: ./gradlew viewpager2:connectedCheck
Change-Id: I8c9b488bb0c49a04dd0a39b2f0d8d7ea66005cdc