Update patch set 1
Patch Set 1:

Please look into this.

I'm submitting this CL to trigger a discussion about the desired behaviour in the case of using same path or path that is a prefix of the other one for assets and resources.

Current implementation allows it and give assets a priority over resources (if a file with the same name exists in resources and assets under the same path, then it'll be opened from assets.).

In my opinion this should not be allowed, resources and assets should be separated similar to having two different directories in the app source (res and assets) and similar to having two different non-overlapping file://android_asset/ and file://android_res/.

My solution is to throw an exception in build() method not set[Assets/Resources]Path methods because while setting paths an overlap can happen I think the validation should be done on the final values. This is for the convenience of the user so that they would not use a temp value or certain order for setting.

{E.g: 
- default path for assets is "/assets/" and for resources is "/res/".
- set resources path to "/assets/andriod_resources/".
- then set assets to "/assets/android_assets/".

the end result would be correct but it would fail if the validation is done in the set method.}

Please let me know your thoughts on this.

Thanks :)

Patch-set: 1
Reviewer: Gerrit User 1141809 <1141809@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1101490 <1101490@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1402701 <1402701@85c56323-6fa9-3386-8a01-6480fb634889>