Update patch set 1
Patch Set 1:

Dear all,

I think that it's OK to keep Context in the UriMediaItem
because UriMediaItem cannot be sent across the process (i.e. cannot be parceled).
When it needs to be sent across the processes, it would be upcasted to MediaItem for sanitizing information.
(It's already documented in UriMediaItem and MediaItem)

It's because the URI may be very private information and doesn't want it to be shared with other apps.
For an example, its URI can points to a load balance server. Media app owner wouldn't want to expose it outside of the app.

Given that, if removing Context is just to make UriMediaItem Parcelable, I rather keep Context there.
What do you think?

Patch-set: 1
CC: Gerrit User 1019013 <1019013@85c56323-6fa9-3386-8a01-6480fb634889>