Separate onBackPressed handling from ComponentActivity
Move the logic in managing OnBackPressedCallbacks
to a separate OnBackPressedDispatcher, deprecating
the related APIs in ComponentActivity in place of
just a getter for the Dispatcher.

Instead of requiring a separate remove method, use
a Subscription which developers can cancel() to remove
the callback.

This allows us to make symmetric APIs for both the
no Lifecycle case and the Lifecycle aware case.

Test: Updated tests
Fixes: 119616705
Change-Id: I780025c3f5e6ff4bb1d55e7569325416a02d8ffa