Fix Robolectric support in workmanager-testing
The issue was introduced (rather counter-intuitively) in this change:
https://android-review.googlesource.com/c/platform/frameworks/support/+/826946/

Before that change, the values across the stack were initialized as follows:
 - initializeTestWorkManager() creates a TestScheduler instance;
 - initializeTestWorkManager() creates a TestWorkManagerImpl instance
    (at this point createSchedulers() is called by the super() constructor);
 - a work is enqueued, getSchedulers() gets called, returning TestScheduler;

After that change, a TestScheduler instance became late-initialized in the
implicit class created by initializeTestWorkManager(), the value was actually
assigned *after* the super() call, which was calling createSchedulers().

This leads to a weird behavior: in the TestWorkManagerImpl instance created by
initializeTestWorkManager, the methods set*Met() work just fine, but in method
createSchedulers() has scheduler == null.

The bottom line is, although this CL should be semantically equivalent, it
actually fixes Robolectric compatibility.

Bug: 125477524
Test: ./gradlew work:work-testing:assemble, run custom Robolectric test
Change-Id: Icb573608b189ff64a4b378a0b1ea78dfd2349893