Update patch set 5
Patch Set 5:

> > The constructor you are using now already creates an in-memory test database anyway. I still think that’s a better approach.
> 
> I meant that the approach I suggested is still a better approach.

I agree that having TestWorkManagerImpl as abstract makes no sense (I've updated the patch to address just that, moving any WorkManagerImpl-related code out).

For the WorkManagerImpl constructor you've mentioned, it's not that simple to "just" supply all the parameters it requires. Let's have a look at what's currently going on in the WorkManagerImpl constructor used at the moment, what we'll have to do ourselves in that case:

        Context applicationContext = context.getApplicationContext();
        WorkDatabase database = WorkDatabase.create(applicationContext, useTestDatabase);
        Logger.setLogger(new Logger.LogcatLogger(configuration.getMinimumLoggingLevel()));
        List<Scheduler> schedulers = createSchedulers(applicationContext);
        Processor processor = new Processor(
                context,
                configuration,
                workTaskExecutor,
                database,
                schedulers);

The worst part here is the dependency of Processor on both database (the code for creation of which we'll have to copy and maintain) *and* the schedulers list *and* the executor: simply supplying the parameters won't work, you need to store them as variables first.
So, basically, WorkManagerTestInitHelper.initializeTestWorkManager() will have to copy all of this code (*bad* idea, as you'll have to have both copies in sync), with the only difference being the way "schedulers" variable is assigned, and then supply the values to the constructor. And then it'll have to still cache the TestScheduler value as a local field to use it as TestDriver delegate and an ExecutionListener.

I rest my case: the current patch is the simplest possible fix for the problem we have, and requires minimum effort to maintain and keep stable, without exposing too much of WorkManagerImpl implementation detail.

Patch-set: 5
Reviewer: Gerrit User 1086529 <1086529@85c56323-6fa9-3386-8a01-6480fb634889>