Update patch set 1
Patch Set 1:

> Discoverability. 

I currently see this API as used mostly by ourselves or people who write some kind of components themselves, so it sound not a lot of traffic. If we are going to suggest to every developer to use it a lot, then I'd agree that we need very strong discoverability and I'd agree to promote it to right in Dispatcher. So do we plan something like that?  

> removeOnBackPressedCallback
Yeah, detach() works in every situation, because we don't expose real backpressure handler,
 and even better, it doesn't require weakreference!

Patch-set: 1
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>