leanback: proper fix for BACK key in SearchEditText
Last CL causes SearchFragment is unable to close when
focus is on EditText because SearchEditText eat the
BACK key in onKeyPreIme.

The original bug root cause: focusing to search result between
onPreIme and OnKeyDown,  because the focus is on result,
the following onKeyDown(BACK) will close the activity.

The fix is to delay the focus to result in post, so that
onKeyDown(BACK) will not close the activity.

Bug: 127221985
Bug: 72306890
Change-Id: Ibbea31b8461abd4dd8cb9a1926b005741a22c76e