Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Gyumin, this is what I've thought. It doesn't look weird nor error-prone.
> What do you think?

This CL would have some bugs if we would not animate the bars when changing size as Jin Seok said in #3 in b/112570875. For example, if users
1. Wait until the bars disappear in FULL size
2. Change the size to EMBEDDED
3. Touch the view to start fade-in animation
They will not be albe to see the CenterView because 2. doesn't bring the bars back to be shown.

Patch-set: 1
Reviewer: Gerrit User 1380121 <1380121@85c56323-6fa9-3386-8a01-6480fb634889>