Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

This LGTM for the purpose of the bug, though this is a bit of a functionality change.

We intentionally did not respect MeasureSpec.EXACTLY bc we wanted apps providing slices to trust that their slice would be displayed in a particular way / nicely.

I think that allowing it to be bigger than ideal height seems OK but I wonder if we should still be enforcing the minimum height requirements for each mode even if EXACTLY is used?

Patch-set: 1
Reviewer: Gerrit User 1033561 <1033561@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2