Fix failing MediaController2Test#test(Set|Adjust)VolumeWithLocalVolume
The minimum alarm volume can be 0 on devices lower than Android P.
However, in the tests the minimum volume was all set to 1 regardless
of the API level.

The tests failed on API < 28 when the devices' current alarm volume
were 0, since they tried to decrease the volume from zero.
This CL fixes this issue by setting the correct minimum volume of the
alarm stream.

Bug: 127589497
Test: Ran MediaController2Test / version-compat-tests (in API 19 and Q)
Change-Id: Iab66b821264884aa2d6b5d5efc6c7fe3111a048c