Update patch set 1
Patch Set 1:

(2 comments)

Thanks a lot for this initiative, really appreciate it! I like the idea of using two well defined methods, instead of one method with arguments that are not very intuitive (and thus make it a weird api).

Two things I would like to do differently, though:
1) I think that the extra layout space in the scroll direction should override the 'general extra layout space', instead of taking the max.
2) I think that the deprecated method should only be called if getExtraLayoutSpaceForScrolling() is not overridden. That is impossible from a language point of view, but we can put up a contract that if a negative value is returned, we fall back to the deprecated method. This is to ensure backwards compatibility: I would like no changes in behavior if someone used the old API and switches to a new version of RV without changing their code.

Patch-set: 1
CC: Gerrit User 1347655 <1347655@85c56323-6fa9-3386-8a01-6480fb634889>