Update patch set 17
Patch Set 17:

(1 comment)

The API LGTM. I'm not that sure about the `currentValue(): T?` of the LiveDataScope<T>. Can you provide some examples of how you'd use it? You said it's useful to recover from a config change. You'd have to implement a state machine or something similar to handle that, right?

Just want to make sure that people don't misuse it and think they're going to get the value of any other LiveData when it's possible they don't. For example, let's say I call `yieldSource(otherLiveData)` and then I call `currentValue()` expecting to get the value that otherLiveData emitted. From my understanding, that's not how the API is supposed to work.

I cannot think of a use case that cannot be solved with this API. Looks amazing!

Thanks,
Manuel

Patch-set: 17
CC: Gerrit User 1451037 <1451037@85c56323-6fa9-3386-8a01-6480fb634889>