Update patch set 1
Patch Set 1:

(2 comments)

> Patch Set 1:
> 
> (2 comments)
> 
> Thanks a lot for this initiative, really appreciate it! I like the idea of using two well defined methods, instead of one method with arguments that are not very intuitive (and thus make it a weird api).
> 
> Two things I would like to do differently, though:
> 1) I think that the extra layout space in the scroll direction should override the 'general extra layout space', instead of taking the max.
> 2) I think that the deprecated method should only be called if getExtraLayoutSpaceForScrolling() is not overridden. That is impossible from a language point of view, but we can put up a contract that if a negative value is returned, we fall back to the deprecated method. This is to ensure backwards compatibility: I would like no changes in behavior if someone used the old API and switches to a new version of RV without changing their code.

Sure thing!

I think you explained it before but why should the extra scroll direction space override the extra layout space?  It seems to me that when a dev wants extra layout space or scroll space, if either is more than what is required, no harm is done, but if one is less than specified, confusing bugs could occur.

I agree with you about the deprecated method.  It's also arguable that I'm assuming too much by removing the RecyclerView.State variable from that method.  I don't see how it is actually useful to that method, but it was put there for a reason.  Probably worth checking in with Yigit about it.  If we brought it back, then the new method could just call the deprecated one, and we would simply be changing the name to something better.

Thoughts on that?

Also, this CL is not fully flushed out... just something I wanted to propose.  I don't plan on following through with this if that is ok with you.  I have enough on my plate.  :)

Patch-set: 1
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>