Update patch set 1
Patch Set 1:

Correction:

> I think there is an inherent ambiguousness when it comes to scroll direction.  When items in the scroll view move toward the top of the scroll view, is the scroll view scrolling up because items are moving in that direction, or down because items further down are becoming progressively exposed...?

Yup, that is an inherent ambiguity indeed. From the perspective of a developer, I think "scrolling towards the end" means "scrolling towards the next page", or similarly "scrolling forward", or similarly and less ambiguously: if the list is in a left-to-right or top-to-bottom context and no reversing of any kind is applied, towards increasing adapter positions. With that definition, mLayoutState.mLastScrollDelta >= 0 gives a scroll direction of TOWARDS_END.

Patch-set: 1