Update patch set 6
Patch Set 6:

I actually used a custom build with this change to write some robolectric tests for our project, and it worked beautifully. I'll try to make a simple test in this repository though (hope it'll work).

As for the weird behavior, I wrote some thoughts on the bug: it seems that robolectric adds some layer of auto-generated classes, which follow the initialization sequence in exactly the same way as the one I've encountered when trying to extract the lambda implicit class into a separate inner class: you can't assign an internal field (in this case it was the TestScheduler) of a subclass before calling super(). With this patch, such behavior won't break us anymore.

Kind regards,

Ivan

Patch-set: 6
Tag: mailMessageId=<CABf=GRdqAJAFXwEYF43pOmOxmuw-JaogKQwz_0hC2Jk20OoxeA@mail.gmail.com>