Update patch set 3
Patch Set 3:

Looks like the docs prebuilts are still looking for the samples in the code in master (which you delete here). This will be fixed once we change the PublishDocsRules to 1.1.0-alpha06, but until that point, perhaps just leave the files referenced in place, but with nothing between the begin/end tags.

Patch-set: 3
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>