Update patch set 2
Patch Set 2: Autosubmit+1 Presubmit-Ready+1

This grammar is good to go.

The various other select rules are not an optimization and are what I can assume to be rules to be consistent with the diagrams in the SQLite website, therefore they are not needed, since the website itself explains the sub-diagrams are just smaller representations of the full select statement diagram.

I checked the various changes with the diagrams and they do look good. Note that this updates the grammar to what was SQLite 3.24.0. The diagrams in the website as of today have two additional changes not present in this change. window functions (https://www.sqlite.org/windowfunctions.html) and enhancements in ALTER TABLE (RENAME COLUMN was added https://www.sqlite.org/lang_altertable.html), both of these grammar changes are in SQLite 3.25.0 which as of today not even Android Q contains (still in 3.22.0).

Patch-set: 2
Reviewer: Gerrit User 1258207 <1258207@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=+1
Label: Presubmit-Ready=+1