Allow extra layout space on both sides
Deprecate the old method getExtraLayoutSpace(state) in favor of a new
mechanism that allows to have extra layout space on both sides, while
keeping the original use case explicit.

The new methods are getExtraLayoutSpaceForScrolling(state) and
getExtraLayoutSpaceAround(state). Backwards compatibility is provided by
calling getExtraLayoutSpace from getExtraLayoutSpaceForScrolling. This
will behave exactly as before if the new methods are not overridden.

Bug: 114217017
Test: ./gradlew recyclerview:connectedCheck
Change-Id: I4a2fc45978ea3304cb9ee581e83739a44ce37a5a