Update patch set 3
Patch Set 3:

(1 comment)

LGTM.

Thanks a lot for taking time to improve the readability of these tests. I was looking into it test right now and I had the same thought that this should be written in a better way, but I couldn't have written it better than this.

Patch-set: 3