Update patch set 1
Patch Set 1:

I'm not sure if I'm particularly happy with this solution (having a second constructor that takes a @LayoutRes), but it does work across library projects and writes well in isolation. It does not compose very well, but perhaps that might be an intentionally good restriction? What do you think?

I'm not entirely sure if it is possible to build tooling to detect if a @ContentView constructor is being called though, I'll chat with the Tools team about that.

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>