Only call applyOverrideConfiguration() if required
Currently we always call applyOverrideConfiguration()
on startup, which means that we create a new Resources
instance, and thus a new cache. This has performance
issues for apps which do not use DayNight.

This CL fixes that by only applying an override
configuration, if the Activity has a different config
to the application.

BUG: 128731115
Test: existing tests pass

Change-Id: I538efde17c35e0d7f2a2e419b9e90a3f2f755ee7