Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-2
> 
> Per offline discussion with lpf@, this is really the wrong layer for handling View events. What was done for Activity (specifically, in cases like startActivity()) and now for Fragment is call view.cancelPendingInputEvents() on the root view, cancelling all pending click events.
> 
> I'd suggest an alternate approach where we cancel pending input events on the dialog's root view as part of DialogFragment as well as do it in the platform in Dialog for Q+ devices.

When the issue happens, we are already inside a click handler. Are you sure that cancelPendingInputEvents inside dismiss would take care of this issue?
Can you provide some more information about the proposed solution?

Just for background, this is an extremely long standing issue (multiple years) and is one of the leading causes of monkey instability. There are 90+ bugs filed on this internally and externally starting in 2011. Unless there are immediate plans to get an alternative fix, I suggest trying to roll out this fix in the meantime (and undoing it once a better solution is ready).

Patch-set: 3
Reviewer: Gerrit User 1149384 <1149384@85c56323-6fa9-3386-8a01-6480fb634889>