Fixed an issue with prematurely removing Fragments
This fixes a theoretical issue when we'd remove a Fragment that is still
attached to another valid ViewHolder.

The current approach (from this patch) would only remove a Fragment if
it is being removed from a ViewHolder it was most recently bound to.
Note: it is valid for a Fragment to be bound to more than one
ViewHolder, and the non-most-recent ViewHolders are considered stale.

Bug: n/a
Test: ./gradlew viewpager2:connectedCheck

Change-Id: I97b758b23fa29ada5858d95fb9d39695e88969ed