Update patch set 3
Patch Set 3: Code-Review-2

mWrap seems to be used to set whether or not the RV's own layout param (in the orientation that scrolling happens) is set to wrap_content or not.  It does seem to be used to set the layout params of children.

Also, where does PagerSnapHelper supply it's own custom child layout params?  I'm not seeing that anywhere.

Patch-set: 3
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2