Limit FragmentScenario#moveToState(STARTED) API to level 24 and above
The timing of onSaveInstanceState callback changed at API level 24 (N) and 28 (P).
Before P, it is called after onPause but before onStop. Any changes made in between
will be lost. In order to workaround this issue, FragmentActivity (to be precise, in
its parent class), marks itself as CREATED state when onSaveInstanceState is called.
This workaround works fine on API level 24 to 27 because it is always followed by
onStop. However, it doesn't work on < 24 (pre-N) because it is called right after
onPause. onStop may not called immediately after causing inconsistent state between
the framework and FragmentActivity.
See http://issuetracker.google.com/65665621#comment3 for additional information.

Unfortunately there is no workaround found that works both pre-N and N. This CL
changes FragmentScenario#moveToState to throw UnsupportedOperationException when
you attempt to move to STARTED state on pre-N devices.

Change-Id: Id3f16510f791accf288e49ee34172c916eeec3f9
Fix: 129002964
Test: fragment-testing