Fix close camera 4 seconds delay in legacy devices
In legacy devices with API level 23 and below, TextureView will release the surfaceTexture when app is in the background.  This will cause camera close
hanging for 4 seconds waiting for the native surface which was released. The solution is to override release() in our own SurfaceTexture class
(FixedSizeSurfaceTexture) and delay the actual release until the SurfaceTexture is totatlly detached (after camera session is closed).

Bug: 128814999
Test: camera.core/CheckedSurfaceTextureTest, FixedSizeSurfaceTextureTest, testapp-core test
Change-Id: I12c7e0e030a02415ecfa5eada2430ac4662ef996