Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Would setHostViewModelStore fit better? That would help tie that this should be set by the NavHost.

> Another option would be to make this just a generic setHost(NavHost) and then instanceof check for ViewModelStoreOwner. That would give us some flexibility in the future for adding Lifecycle, SavedState, etc. from the NavHost without requiring additional work on the part of custom NavHosts. WDYT?

For option1: sounds ok for me.  

For option2: I don't really like that we cast internally, so it hides a dependency. If we introduce interface that combines SavedStateRegistryOwner (it already has LifecycleOwner) and  ViewModelStoreOwner and just receive it as a parameter than it would be ideal. For now, we can do that ugly trick with generic (T extends SavedStateRegistryOwner & ViewModelStoreOwner)

Also it feels like it should be in constructor, because now it is required to be called during setup and otherwise NavController isn't fully functioning. Then we deprecate current one and introduce NavContoller(Context, ComponentInterface) 
  
Similarly, what do you think? :)

Patch-set: 1
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0