Update patch set 1
Patch Set 1:

> I think there will be NavHosts that won't have access to a ViewModelStoreOwner (say, a View based one) and that making it optional still makes sense.


Yeah, it is a bit of the issue then too, because once you get a NavController object there is no way to check if it supports ViewModels or not? While it is one-off method it is not a big deal, but if you plan to add other methods like that, then it would be better to do something like: navController.componentExtensions()?.getViewModelStore(id). This componentExtensions would have all these methods that don't work with view but work if you provide viewmodelStore & co  
 
> The big problem with ComponentInterface is that we can never change the interface since it would break binary compatibility (we couldn't have it extend only ViewModelStoreOwner and then later add a requirement that it also extend SavedStateRegistryOwner without breaking existing libraries).

Yeah, I know =|

Patch-set: 1
Reviewer: Gerrit User 1098853 <1098853@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0