Update patch set 5
Patch Set 5:

(6 comments)

This is not something I expect you to change in this cl, but maybe something to consider in the future.

I'm not a big fan of condensing all of the tested things into one parameterized test (which then requires pushing code paths through many if statements).  I recognize that this strategy results in less code.... but I think that comes at the cost of less clarity in the test, and I think clarity wins when a test fails and some other reader has to come along and understand why (or modify a test without fear of breaking it).

Just something to consider (and possibly play with) in terms of test writing strategy in the future.  I might also be in a minority here... :)

That being said, I think the test looks good!  I did have one comment about the actual verification you are doing (which seems to allow for the possibility of not noticing bad behavior that the new code could create), and then otherwise nits that you can take or leave.

Thanks for all of the hard work!

Patch-set: 5
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>