Update patch set 2
Patch Set 2:

This is removing the setHwLayerAnimationListenerIfAlpha wrapper animation, which is necessary - why is this particular case a problem when this same pattern is used in many other places? There's still only one animation being passed to startAnimation, which seems to contradict the commit message?

Patch-set: 2
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>