Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> This is removing the setHwLayerAnimationListenerIfAlpha wrapper animation, which is necessary - why is this particular case a problem when this same pattern is used in many other places? There's still only one animation being passed to startAnimation, which seems to contradict the commit message?

Not entirely sure why this one is the culprit. I was hoping someone might have more context surrounding that. It seems to not cause issues in all of the other places. The Animation produced by setHwLayerAnimationListenerIfAlpha is exactly the same as the one passed in, it just stores the original animation in a list. 

A "single" animation was bad wording, unnested or unwrapped would be better.

Could there be some weird underlying moveState condition that makes this use different from the others?

Patch-set: 2
Reviewer: Gerrit User 1442279 <1442279@85c56323-6fa9-3386-8a01-6480fb634889>