On-demand initialization for WorkManager.
This CL introduces WorkManager#getInstance(Context) as the preferred
method and deprecates the argument-less version.  This method performs
on-demand initialization of WorkManager if it hasn't been initialized
previously by looking to see if the application Context implements
Configuration.Provider and querying it for the Configuration.

The exact order of initialization checks are now as follows:

1. Through WorkManagerInitializer if it's enabled.
2. Otherwise, through a WorkManager#initialize in Application#onCreate
   or ContentProvider#onCreate if specified by the developer.
3. Otherwise, through WorkManager#getInstance(Context) if the
   application Context implements Configuration.Provider.  Using
   WorkManager#getInstance() at this point results in an exception as
   it does not contain the necessary information to initialize.
4. Otherwise, throws an IllegalStateException.

This change allows WorkManager to be setup in a manner where you can
only allow it to be initialized when it is needed.  This is good for
apps that care about the critical path initialization needed by
WorkManager; in addition, all of WorkManager's own BroadcastReceivers
and Services are set up to try to auto-initialize if needed.  Note
that this initialization can still happen on the main thread.

Bug: 127497100
Test: Ran integration tests with integration app doing on-demand init.
Change-Id: I30930d6445c0e9c5f08eaaab554963aaac99b3c9