Update patch set 1
Patch Set 1:

If the problem is that we're not updating the state of postponed Fragments, it seems like updating moveToState(int newState, boolean always) to also iterate over postponed Fragments would also solve this issue (and ensure that postponed Fragments are stopped, etc. when those events happen). Was there a reason this approach was used over that?

Patch-set: 1
Reviewer: Gerrit User 1060451 <1060451@85c56323-6fa9-3386-8a01-6480fb634889>