Manually perform SwipeToLocation on API 15
On API 16 and above, it is possible to execute Espresso's ViewAction's
during a smooth scroll of a RecyclerView, because Espresso doesn't
consider the scroll as 'non-idle'. On API 15, however, Espresso will
delay any ViewAction because it considers the UI thread to be non-idle.

Solve this by manually executing the ViewAction on API 15, instead of
relying on Espresso's mechanisms. This requires a slight refactor in
SwipeToLocation, because we don't have a UiController at our disposal.
Instead, we have to inject motion events with an Instrumentation object.

Bug: 128859160
Test: ./gradlew viewpager2:connectedCheck
Change-Id: I5bd68ba1bee228f6d0c162024f41488d196f84d3