Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If the problem is that we're not updating the state of postponed Fragments, it seems like updating moveToState(int newState, boolean always) to also iterate over postponed Fragments would also solve this issue (and ensure that postponed Fragments are stopped, etc. when those events happen). Was there a reason this approach was used over that?

Yes, I did it by making the iteration over active fragments in moveToState() include postponed fragments as well. That requires that moveFragmentToExpectedState() also accounts for postponed fragments. The problem is that a postponed fragment is also removed. So if a user postponed a fragment and then popped it from the back stack without ever actually executing the animation, we get a regression. There is no distinction between a fragment that is just postponed and one that actually needs to be removed.

Patch-set: 1
Reviewer: Gerrit User 1442279 <1442279@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0